requirements.txt
APScheduler
python-dotenv
psycopg2-binary
aiogram